- fix swimlane - remove boolean from logout and add mate - correct return type for getUserInput functions - refactor ProjectsDashboardView - refactor ProjectDetailsView - renamed password and userName validators - fixed a csv file bug - remove unnecessary files - refactor login view - refactor main menu view - add boolean return to ViewExceptionHandler

+++ src/main/kotlin/ui/taskManagementView/TaskStateEditionView.kt
@@ -21,7 +21,7 @@ class TaskStateEditionView(
             return
         }
         printProjectState(projectTasksStates)
-        val newStateIndex = cliReader.getValidUserNumberInRange(min = 1, max = projectTasksStates.size).toInt() - 1
+        val newStateIndex = cliReader.getValidUserNumberInRange(min = 1, max = projectTasksStates.size) - 1
 
         viewExceptionHandler.tryCall {
             manageTaskUseCase.editTaskState(task.id, projectTasksStates[newStateIndex])
